poseflow/backend/requirements.txt
# PoseFlow Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.26.3
Pillow==10.2.0
